review changes + inpovements
+ review changes von andx
+ schnee wird höher erstellt damit dieser so wenig wie möglich im sichtfeld spawnt
+ exit If Verbesserungen
