+++ addons/effects/config.cpp
@@ -14,47 +14,5 @@ class CfgPatches {
     };
 };
 
-class CfgSounds {
-    class wind_1 {
-        name = "Wind 1";
-        sound[] = {"a3\data_f_curator\sound\cfgsounds\wind1.wss", 1, 1};
-        titles[] = {};
-    };
-    class wind_2 {
-        name = "Wind 2";
-        sound[] = {"a3\data_f_curator\sound\cfgsounds\wind2.wss", 1, 1};
-        titles[] = {};
-    };
-    class wind_3 {
-        name = "Wind 3";
-        sound[] = {"a3\data_f_curator\sound\cfgsounds\wind3.wss", 1, 1};
-        titles[] = {};
-    };
-    class wind_4 {
-        name = "Wind 4";
-        sound[] = {"a3\data_f_curator\sound\cfgsounds\wind4.wss", 1, 1};
-        titles[] = {};
-    };
-    class wind_5 {
-        name = "Wind 5";
-        sound[] = {"a3\data_f_curator\sound\cfgsounds\wind5.wss", 1, 1};
-        titles[] = {};
-    };
-    class wind_big {
-        name = "Wind Big";
-        sound[] = {"a3\missions_f_bootcamp\data\sounds\wind_big.wss", 1, 1};
-        titles[] = {};
-    };
-    class wind_normal {
-        name = "Wind Normal";
-        sound[] = {"a3\missions_f_bootcamp\data\sounds\wind_normal.wss", 1, 1};
-        titles[] = {};
-    };
-    class wind_small {
-        name = "Wind Small";
-        sound[] = {"a3\missions_f_bootcamp\data\sounds\wind_small.wss", 1, 1, 25 };
-        titles[] = {};
-    };
-};
-
 #include "CfgEventHandlers.hpp"
+#include "CfgSounds.hpp"
