+++ addons/effects/script_component.hpp
@@ -1,5 +1,5 @@
 #define COMPONENT effects
-#define COMPONENT_BEAUTIFIED TTT Effekte
+#define COMPONENT_BEAUTIFIED Effekte
 #include "\z\ttt\addons\main\script_mod.hpp"
 
 // #define DEBUG_MODE_FULL
